+++ src/data/dummyData.js
@@ -38,31 +38,43 @@ export const data = {
     "justify-content": "justify-content",
     "align-item": "align-item",
   },
-  js1: {
-    declare: "선언",
-    assign: "할당",
-    declare2: "선언",
-    assign2: "할당",
-    typeof: "typeof",
-    string: "string",
-    number: "number",
-    boolean: "boolean",
-    undefined: "undefined",
-    function: "function",
-    object: "object",
-    object2: "object",
-    func: "함수",
-    declareStyle: "선언식",
-    expressionStyle: "표현식",
-    arrow: "화살표",
-    not: "!",
-    and: "&&",
-    or: "||",
-    upper: "toUpperCase()",
-    trim: "trim()",
-    split: "split()",
-    reset: "초기화",
-  },
+  js1: [
+    {
+      declare: "선언",
+      assign: "할당",
+      declare2: "선언",
+      assign2: "할당",
+    },
+    {
+      typeof: "typeof",
+      string: "string",
+      number: "number",
+      boolean: "boolean",
+      undefined: "undefined",
+      function: "function",
+      object: "object",
+      object2: "object",
+    },
+    {
+      func: "함수",
+      declareStyle: "선언식",
+      expressionStyle: "표현식",
+      arrow: "화살표",
+    },
+    {
+      not: "!",
+      and: "&&",
+      or: "||",
+    },
+    {
+      upper: "toUpperCase()",
+      trim: "trim()",
+      split: "split()",
+    },
+    {
+      reset: "초기화",
+    },
+  ],
   js2: {},
   test: [
     {
